+++ FluidNC/src/ControlPin.h
@@ -3,14 +3,18 @@
 #include "Pin.h"
 #include "Machine/EventPin.h"
 namespace Machine {
-    class ControlPin : public Machine::EventPin {
+    class ControlPin : public EventPin {
     private:
         const char _letter;  // The name that appears in init() messages and the name of the configuration item
 
     public:
-        ControlPin(Event* event, const char* legend, char letter) : EventPin(event, legend, &_pin), _letter(letter) {}
+        ControlPin(Event* event, const char* legend, char letter) : EventPin(event, legend), _letter(letter) {}
+
+        void init();
 
         Pin _pin;
+
+        bool get() { return _pin.read(); }
 
         char letter() { return _letter; };
