+++ FluidNC/src/Pins/PinDetail.cpp
@@ -7,13 +7,8 @@
 #include <esp_attr.h>  // IRAM_ATTR
 
 namespace Pins {
-    void PinDetail::attachInterrupt(void (*callback)(void*), void* arg, int mode) {
-        Assert(false, "Interrupts are not supported by pin %d", _index);
-    }
-    void PinDetail::detachInterrupt() {
-        Assert(false, "Interrupts are not supported by pin %d", _index);
-        ;
-    }
+    void PinDetail::registerEvent(EventPin* obj) { Assert(false, "registerEvent is not supported by pin %d", _index); }
+
     void IRAM_ATTR PinDetail::synchronousWrite(int high) { write(high); }
 
 }
