+++ FluidNC/test/Pins/ErrorPinTest.cpp
@@ -16,9 +16,6 @@ namespace Pins {
         AssertThrow(errorPin.write(true));
         AssertThrow(errorPin.read());
 
-        AssertThrow(errorPin.attachInterrupt([](void* arg) {}, EITHER_EDGE));
-        AssertThrow(errorPin.detachInterrupt());
-
         Assert(errorPin.capabilities() == Pin::Capabilities::Error, "Incorrect caps");
     }
 }
